renaming config and args members
renamed port to upstream port for easier understanding
